Squashed commit of the following:
commit 5eb1d49c55e0c9c1d2c9a414b923d19dae3f2a30
Author: Julian Callin <julian.callin@here.com>
Date:   Fri Jan 4 10:24:15 2019 -0800

    Update readme with array algorithms

commit 33be08e1925faa2b413058cd6b61e4ad6e1c05d9
Author: Julian Callin <julian.callin@here.com>
Date:   Fri Jan 4 10:17:27 2019 -0800

    Move all array algorithms to their own objects
