+++ README.md
@@ -7,6 +7,12 @@
   - [Singly Linked List](data-structures/src/main/scala/com/scalaprojects/datastructures/linkedlists/SinglyLinkedList.scala)
 ## Algorithms
 * [Array Algorithms](algorithms/src/main/scala/com/scalaprojects/algorithms/Arrays.scala)
+  - [Find adjacent pairs which sum to n](algorithms/src/main/scala/com/scalaprojects/algorithms/arrays/FindAdjacentPairsWhichSumToN.scala)
+  - [Find the missing number in a range](algorithms/src/main/scala/com/scalaprojects/algorithms/arrays/FindMissingNumberInRange.scala)
+  - [Remove duplicates](algorithms/src/main/scala/com/scalaprojects/algorithms/arrays/RemoveDuplicates.scala)
+  - [Reverse array iteratively](algorithms/src/main/scala/com/scalaprojects/algorithms/arrays/ReverseArrayIterative.scala)
+  - [Reverse array recursively](algorithms/src/main/scala/com/scalaprojects/algorithms/arrays/ReverseArrayRecursive.scala)
+  - [Swap items](algorithms/src/main/scala/com/scalaprojects/algorithms/arrays/SwapItems.scala)
 * Sorting
   - [Quicksort (Simplified)](algorithms/src/main/scala/com/scalaprojects/algorithms/sorting/QuickSortSimplified.scala)
   - [Quicksort](algorithms/src/main/scala/com/scalaprojects/algorithms/sorting/QuickSort.scala)
