+++ algorithms/src/main/scala/com/scalaprojects/algorithms/sorting/QuickSortSimplified.scala
@@ -5,17 +5,17 @@ object QuickSortSimplified {
   type IntList = List[Int]
 
   /**
-    * Sort an array by quicksort using some scala sugar and List instead of Array
-    *
-    * Time:
-    * worst:   n2 (randomize partition index for nlogn average)
-    * average: nlogn
-    * best:    n
-    *
-    * Space:   n (in place)
-    *
-    * @param l list to sort
-    */
+   * Sort an array by quicksort using some scala sugar and List instead of Array
+   *
+   * Time:
+   * worst:   n2 (randomize partition index for nlogn average)
+   * average: nlogn
+   * best:    n
+   *
+   * Space:   n (in place)
+   *
+   * @param l list to sort
+   */
   def quickSort(l: IntList): IntList = {
     def quickSortHelper(l: IntList): IntList = {
       l match {
@@ -27,11 +27,11 @@ object QuickSortSimplified {
     }
 
     /**
-      * Return partitions of a list partitioned around the first element
-      *
-      * @param l list to partition
-      * @return (pivot value, items lesser than pivot, items greater than pivot)
-      */
+     * Return partitions of a list partitioned around the first element
+     *
+     * @param l list to partition
+     * @return (pivot value, items lesser than pivot, items greater than pivot)
+     */
     def partition(l: IntList): (Int, IntList, IntList) = {
       val pivot :: rest = l
       val (lesser, greater) = rest.partition(_ < pivot)
